Support Tornado6
This commit adds support for Tornado 6. Only change users are expected
to make is to import TornadoScopeManager from
`tornado_opentracing.scope_managers` instead of
`opentracing.scope_managers.tornado`.

This is required as tornado_opentracing automatically exports the
correct scope manager to use depending on the version of Python and
Tornado being used.