Bootstrap command can now optionally print deps
Updated bootstrap command to allow users to optionally print deps
instead of installing them.

This add a new flag `-r` or `--requirements` to the bootstrap command.
When set, it forces the command to print the list of dependencies it
would've otherwise installed to console instead of actually installing
them.

This is useful for environments where the bootstrap command can't be
executed easily for example AWS Lambda.