requests: support sub-classed sessions
We patch the Session object such that it actually returns instances
of our subclass (SessionTracing). This breaks when users use a sub-class
of requests.Session as our implementation subclasses requests.Session
directly.

For example, when using requests-oauthlib, none of the methods provided
by requests_oauthlib.Session can be accessed as we actually force create
SessionTracing instances insted of requests_oauthlib.Session.

This commmit checks whether the user is requesting an instance of
requests.Session or a subclass of it. If it requests, it preserves the
current behavior. If it is a subclass, it dynamically creates another
class that derives from the subclass and SessionTracing.